split confgen into MVC pattern
